+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import FlexLayout from "./components/layout/FlexLayout";
-import MainSection from "./components/section/MainSection";
-import NavSection from "./components/section/NavSection";
+import MainSection from "./components/layout/MainSection";
+import NavSection from "./components/layout/NavSection";
 
 const App = () => {
   return (
