+++ src/components/nav/NavNoteListItem.jsx
