+++ src/components/notes/NoteArea.jsx
