+++ src/components/notes/NoteTextArea.jsx
