+++ src/components/routes/Notes.jsx
@@ -1,6 +1,6 @@
-import NoteList from "../NoteList";
-import NoteArea from "../ui/NoteArea";
-import SortedArea from "../ui/SortedArea";
+import NoteList from "../notes/NoteList";
+import NoteArea from "../notes/NoteArea";
+import SortedArea from "../notes/SortedArea";
 
 const Notes = () => {
   return (
